marker InfoWindow set with custom text using LandmarkEntityObject. HashMaps no longer needed to get reference to LandmarkEntity - thanks google. Next: add thumbnail and description to info window.
