+++ app/src/main/java/rocks/morrisontech/historicsf/OnTaskCompleted.java
@@ -1,0 +1,11 @@
+package rocks.morrisontech.historicsf;
+
+import rocks.morrisontech.historicsf.entity.LandmarkEntity;
+
+/**
+ * Created by Quinn on 1/7/17.
+ */
+
+public interface OnTaskCompleted {
+    void onTaskCompleted(LandmarkEntity[] entities);
+}
