+++ app/src/main/res/layout/marker_info.xml
@@ -3,12 +3,14 @@
     android:layout_width="wrap_content" android:layout_height="wrap_content">
 
     <TextView
+        android:id="@+id/title_textView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Info Window"
          />
 
     <ImageView
+        android:id="@+id/thumbnail_imageView"
         android:layout_width="16dp"
         android:layout_height="16dp"
         />
